% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a map for the location of accidents within a state for a given year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{(integer) the number of the ID of the state to analyse.

This function depends on the dplyr::filter, maps::map functions.
bind_rows, group_by, and summarize. It also depends on tidyr::spread.}

\item{year}{(integer or character) the year identifying the file whose name
to construct.}
}
\value{
NULL. Function has side effect of plotting a map.
}
\description{
Create a map for the location of accidents within a state for a given year.
}
