% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV data file as a data frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{(character) the name of the file to be read.}
}
\value{
(data frame, tbl_df, tbl) data read from the file
}
\description{
The filename must refer to an exsting filename containing comma-separated
data.
This function depends on the readr::read_csv function.
}
