% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read and combine data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{(integer or character) vector or list of the years for which to
read data.

The `years` parameter must be supplied in a form which can be converted to
integer year value (by `as.integer`).

This function depends on dplyr::select and dplyr::mutate.}
}
\value{
(list) of data frames each entry having data for a specific year.
}
\description{
Read and combine data for multiple years
}
