% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a data file name for a given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{(integer or character) the year identifying the file whose name
  to construct.

The "year" parameter must be supplied in a form which can be converted to
an integer year value (by `as.integer`).}
}
\value{
(character) file name incorporating year.
}
\description{
Create a data file name for a given year.
}
\examples{
\dontrun{
  make_filename(2017)
  make_filename("2017")
}
}
