% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data read for multiple years by number of accidents per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{(integer or character) vector or list of the years for which to
read data.

The `years` parameter must be supplied in a form which can be converted to
integer year value (by `as.integer`).

This function depends on several functions from the dplyr package:
bind_rows, group_by, and summarize. It also depends on tidyr::spread.}
}
\value{
(data frame, tbl_df, tbl) a data frame containing the accident count of every year selected (one column per year) computed by month (one row per month)
}
\description{
Summarize data read for multiple years by number of accidents per month
}
